refactor(structure): align folders with target organization
