--- CruiserJumpPractice/Application/ClientOperationResults.cs
@@ -1,16 +1,0 @@
-#nullable enable
-
-namespace CruiserJumpPractice.Application;
-
-internal enum HostGuardResult
-{
-    Success,
-    HostOnly
-}
-
-internal enum ToggleMagnetResult
-{
-    HostOnly,
-    MagnetOn,
-    MagnetOff
-}
