--- CruiserJumpPractice/Application/CruiserStateOperationResults.cs
@@ -1,19 +1,0 @@
-#nullable enable
-
-namespace CruiserJumpPractice.Application;
-
-internal enum SaveCruiserStateResult
-{
-    Success,
-    NoCruiserFound,
-    UnexpectedState
-}
-
-internal enum LoadCruiserStateResult
-{
-    Success,
-    NoCruiserFound,
-    NoSavedState,
-    MagnetedToShip,
-    UnexpectedState
-}
