--- CruiserJumpPractice/Application/Services/ServerCruiserStateService.cs
@@ -1,30 +1,0 @@
-#nullable enable
-
-using CruiserJumpPractice.Application.UseCases;
-
-namespace CruiserJumpPractice.Application.Services;
-
-internal sealed class ServerCruiserStateService
-{
-    private readonly SaveCruiserStateUseCase saveCruiserStateUseCase;
-    private readonly LoadCruiserStateUseCase loadCruiserStateUseCase;
-
-    public ServerCruiserStateService(
-        SaveCruiserStateUseCase saveCruiserStateUseCase,
-        LoadCruiserStateUseCase loadCruiserStateUseCase
-    )
-    {
-        this.saveCruiserStateUseCase = saveCruiserStateUseCase;
-        this.loadCruiserStateUseCase = loadCruiserStateUseCase;
-    }
-
-    internal SaveCruiserStateResult SaveCruiserState()
-    {
-        return saveCruiserStateUseCase.Execute();
-    }
-
-    internal LoadCruiserStateResult LoadCruiserState()
-    {
-        return loadCruiserStateUseCase.Execute();
-    }
-}
