--- CruiserJumpPractice/Application/UseCases/RequestSaveCruiserStateUseCase.cs
@@ -1,27 +1,0 @@
-#nullable enable
-
-using CruiserJumpPractice.GameInterop;
-
-namespace CruiserJumpPractice.Application.UseCases;
-
-internal sealed class RequestSaveCruiserStateUseCase
-{
-    private readonly IGameInterop gameInterop;
-
-    public RequestSaveCruiserStateUseCase(IGameInterop gameInterop)
-    {
-        this.gameInterop = gameInterop;
-    }
-
-    public HostGuardResult Execute()
-    {
-        if (!gameInterop.IsHost())
-        {
-            return HostGuardResult.HostOnly;
-        }
-
-        var rpcSurrogateNetworkBehaviour = gameInterop.GetRpcSurrogateNetworkBehaviour();
-        rpcSurrogateNetworkBehaviour.SaveCruiserStateServerRpc();
-        return HostGuardResult.Success;
-    }
-}
