+++ CruiserJumpPractice/Domain/CruiserStateResults.cs
@@ -1,0 +1,19 @@
+#nullable enable
+
+namespace CruiserJumpPractice.Domain;
+
+internal enum SaveCruiserStateResult
+{
+    Success,
+    NoCruiserFound,
+    UnexpectedState
+}
+
+internal enum LoadCruiserStateResult
+{
+    Success,
+    NoCruiserFound,
+    NoSavedState,
+    MagnetedToShip,
+    UnexpectedState
+}
