+++ CruiserJumpPractice/Domain/MagnetResults.cs
@@ -1,0 +1,16 @@
+#nullable enable
+
+namespace CruiserJumpPractice.Domain;
+
+internal enum HostGuardResult
+{
+    Success,
+    HostOnly
+}
+
+internal enum ToggleMagnetResult
+{
+    HostOnly,
+    MagnetOn,
+    MagnetOff
+}
