--- CruiserJumpPractice/GameInterop/GameInteropException.cs
@@ -1,10 +1,0 @@
-#nullable enable
-
-namespace CruiserJumpPractice.GameInterop;
-
-internal sealed class GameInteropException : System.Exception
-{
-    public GameInteropException(string message) : base(message)
-    {
-    }
-}
