+++ CruiserJumpPractice/Interop/Domain/GameInteropException.cs
@@ -1,0 +1,10 @@
+#nullable enable
+
+namespace CruiserJumpPractice.Interop.Domain;
+
+internal sealed class GameInteropException : System.Exception
+{
+    public GameInteropException(string message) : base(message)
+    {
+    }
+}
