+++ CruiserJumpPractice/Interop/IGameInterop.cs
@@ -1,0 +1,37 @@
+#nullable enable
+
+using CruiserJumpPractice.Interop.Behaviours;
+using CruiserJumpPractice.Interop.Domain;
+
+namespace CruiserJumpPractice.Interop;
+
+internal interface IGameInterop
+{
+    bool IsServer();
+
+    bool IsClient();
+
+    bool IsHost();
+
+    bool IsLocalPlayerBusy();
+
+    void DisplayTip(HUDManager hudManager, string headerText, string bodyText);
+
+    void DisplayLocalTip(string headerText, string bodyText);
+
+    void SpawnRpcSurrogate();
+
+    RpcSurrogateNetworkBehaviour GetRpcSurrogateNetworkBehaviour();
+
+    VehicleController? FindCruiser();
+
+    CruiserSnapshot CaptureCruiser(VehicleController cruiser);
+
+    void RestoreCruiser(VehicleController cruiser, CruiserSnapshot snapshot);
+
+    bool IsCruiserMagnetedToShip(VehicleController cruiser);
+
+    bool IsShipMagnetOn();
+
+    void ToggleShipMagnet();
+}
