--- CruiserJumpPractice/Presentation/ClientCruiserStateService.cs
@@ -1,42 +1,0 @@
-#nullable enable
-
-using CruiserJumpPractice.Application;
-using CruiserJumpPractice.Application.UseCases;
-
-namespace CruiserJumpPractice.Presentation;
-
-internal sealed class ClientCruiserStateService
-{
-    private readonly RequestSaveCruiserStateUseCase requestSaveCruiserStateUseCase;
-    private readonly RequestLoadCruiserStateUseCase requestLoadCruiserStateUseCase;
-    private readonly ClientNotificationService clientNotificationService;
-
-    public ClientCruiserStateService(
-        RequestSaveCruiserStateUseCase requestSaveCruiserStateUseCase,
-        RequestLoadCruiserStateUseCase requestLoadCruiserStateUseCase,
-        ClientNotificationService clientNotificationService
-    )
-    {
-        this.requestSaveCruiserStateUseCase = requestSaveCruiserStateUseCase;
-        this.requestLoadCruiserStateUseCase = requestLoadCruiserStateUseCase;
-        this.clientNotificationService = clientNotificationService;
-    }
-
-    internal void RequestSaveCruiserState()
-    {
-        var result = requestSaveCruiserStateUseCase.Execute();
-        if (result == HostGuardResult.HostOnly)
-        {
-            clientNotificationService.ShowCruiserTip("Only the host can save the cruiser state.");
-        }
-    }
-
-    internal void RequestLoadCruiserState()
-    {
-        var result = requestLoadCruiserStateUseCase.Execute();
-        if (result == HostGuardResult.HostOnly)
-        {
-            clientNotificationService.ShowCruiserTip("Only the host can load the cruiser state.");
-        }
-    }
-}
