+++ CruiserJumpPractice/Runtime/StartupHandler.cs
@@ -1,0 +1,20 @@
+#nullable enable
+
+using CruiserJumpPractice.Interop;
+
+namespace CruiserJumpPractice.Runtime;
+
+internal sealed class StartupHandler
+{
+    private readonly IGameInterop gameInterop;
+
+    public StartupHandler(IGameInterop gameInterop)
+    {
+        this.gameInterop = gameInterop;
+    }
+
+    public void HandleStartup()
+    {
+        gameInterop.SpawnRpcSurrogate();
+    }
+}
