+++ CruiserJumpPractice/Services/ServerCruiserStateService.cs
@@ -1,0 +1,31 @@
+#nullable enable
+
+using CruiserJumpPractice.Domain;
+using CruiserJumpPractice.UseCases;
+
+namespace CruiserJumpPractice.Services;
+
+internal sealed class ServerCruiserStateService
+{
+    private readonly SaveCruiserStateUseCase saveCruiserStateUseCase;
+    private readonly LoadCruiserStateUseCase loadCruiserStateUseCase;
+
+    public ServerCruiserStateService(
+        SaveCruiserStateUseCase saveCruiserStateUseCase,
+        LoadCruiserStateUseCase loadCruiserStateUseCase
+    )
+    {
+        this.saveCruiserStateUseCase = saveCruiserStateUseCase;
+        this.loadCruiserStateUseCase = loadCruiserStateUseCase;
+    }
+
+    internal SaveCruiserStateResult SaveCruiserState()
+    {
+        return saveCruiserStateUseCase.Execute();
+    }
+
+    internal LoadCruiserStateResult LoadCruiserState()
+    {
+        return loadCruiserStateUseCase.Execute();
+    }
+}
