+++ CruiserJumpPractice/UseCases/RequestLoadCruiserStateUseCase.cs
@@ -1,0 +1,28 @@
+#nullable enable
+
+using CruiserJumpPractice.Interop;
+using CruiserJumpPractice.Domain;
+
+namespace CruiserJumpPractice.UseCases;
+
+internal sealed class RequestLoadCruiserStateUseCase
+{
+    private readonly IGameInterop gameInterop;
+
+    public RequestLoadCruiserStateUseCase(IGameInterop gameInterop)
+    {
+        this.gameInterop = gameInterop;
+    }
+
+    public HostGuardResult Execute()
+    {
+        if (!gameInterop.IsHost())
+        {
+            return HostGuardResult.HostOnly;
+        }
+
+        var rpcSurrogateNetworkBehaviour = gameInterop.GetRpcSurrogateNetworkBehaviour();
+        rpcSurrogateNetworkBehaviour.LoadCruiserStateServerRpc();
+        return HostGuardResult.Success;
+    }
+}
